# chess_game/internal/api/auth.go
package api

import (
	"errors"
	"log/slog"
	"net/http"
	"strings"
	"time"

	db "github.com/Glenn444/golang-chess/internal/db"
	"github.com/Glenn444/golang-chess/internal/token"
	"github.com/Glenn444/golang-chess/internal/utils/auth"
	"github.com/gin-gonic/gin"
	"github.com/jackc/pgx/v5"
	"github.com/jackc/pgx/v5/pgtype"
)

// createUser (registration IS an auth concern)
// confirmEmail
// loginUser
// refreshToken

type CreateUserRequest struct {
	Username string `json:"username" binding:"required,alphanum"`
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=6"`
}

func (r *CreateUserRequest) SanitizeCreateUserReq() {
	r.Username = strings.ToLower(r.Username)
	r.Email = strings.ToLower(r.Email)
}

type CreateUserResponse struct {
	Username          string    `json:"username"`
	Email             string    `json:"email"`
	PasswordChangedAt time.Time `json:"password_changed_at"`
	CreatedAt         time.Time `json:"created_at"`
}

// @Summary      Register a new user
// @Description  Creates an account and sends a 6-digit OTP to the provided email.
// @Tags         Auth
// @Accept       json
// @Produce      json
// @Param        body  body  CreateUserRequest  true  "Registration payload"
// @Success      200   {object}  CreateUserResponse
// @Failure      400   {object}  map[string]string
// @Failure      409   {object}  map[string]string
// @Failure      500   {object}  map[string]string
// @Router       /users/signup [post]
func (server *Server) createUser(ctx *gin.Context) {

	var req CreateUserRequest
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, errorResponse(err))
		return
	}

	//sanitize user input
	req.SanitizeCreateUserReq()

	//check if username exists, ignore if does not exist
	userExists, err := server.store.UsernameExists(ctx, req.Username)
	if handleDBError(ctx, err,
		WithLogArgs("createUser: failed to check UserNameExists", "username", req.Username),
	) {
		return
	}

	if userExists {
		ctx.JSON(http.StatusConflict, errorMessage(ErrUserAlreadyExists))
		return
	}

	hashedPassword, err := auth.HashPassword(req.Password)
	if err != nil {
		slog.Error("createUser: failed to hash Password", "err", err, "user_email", req.Email)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrInternalServer))
		return
	}

	arg := db.CreateUserParams{
		Username:     req.Username,
		Email:        req.Email,
		PasswordHash: hashedPassword,
	}

	user, err := server.store.CreateUser(ctx, arg)
	if handleDBError(ctx, err,
		WithUniqueMsg(ErrUserAlreadyExists),
		WithLogArgs("createUser: failed creating user in db ", "user_email", req.Email)) {
		return
	}

	otpCode, err := auth.GenerateOTP(6)
	if err != nil {
		slog.Error("createrUser: failed to generate OTP Code", "err", err, "user_id", user.ID)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrGeneratingOTP))
		return
	}
	otpHash, err := auth.SignOtpCode(otpCode, server.config.TokenSymmetricKey)
	if err != nil {
		slog.Error("createUser: failed SignOtpCode", "err", err, "user_id", user.ID)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrInternalServer))
		return
	}

	argEmailOtp := db.CreateEmailOTPParams{
		UserID:    user.ID,
		CodeHash:  otpHash,
		ExpiresAt: pgtype.Timestamptz{Time: time.Now().Add(server.config.AcessTokenDuration), Valid: true},
	}
	//verify email
	_, err = server.store.CreateEmailOTP(ctx, argEmailOtp)
	if handleDBError(ctx, err,
		WithLogArgs("createUser: failed CreateEmailOTP", "user_id", user.ID),
	) {
		return
	}

	//send the otp to the user email
	err = server.emailClient.SendEmailOTP(user.Email, otpCode)
	if err != nil {
		slog.Error("createUser: failed SendEmailOTP", "err", err, "user_email", user.Email)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrSendingOTP))
		return
	}
	resp := CreateUserResponse{
		Username:          user.Username,
		PasswordChangedAt: user.UpdatedAt.Time,
		CreatedAt:         user.CreatedAt.Time,
	}
	ctx.JSON(http.StatusOK, resp)

}

type ConfirmEmail struct {
	Email     string `json:"email" binding:"required,email"`
	EMAIL_OTP string `json:"email_otp" binding:"required,numeric,len=6"`
}

func (r *ConfirmEmail) SanitizeConfirmEmailReq() {
	r.Email = strings.ToLower(r.Email)
}

type ConfirmEmailResponse struct {
	Message string `json:"message"`
	Email   string `json:"email"`
}

// @Summary      Confirm email
// @Description  Verifies email address with the 6-digit OTP code.
// @Tags         Auth
// @Accept       json
// @Produce      json
// @Param        body  body  ConfirmEmail  true  "OTP payload"
// @Success      200   {object}  ConfirmEmailResponse
// @Failure      400   {object}  map[string]string
// @Failure      403   {object}  map[string]string
// @Failure      404   {object}  map[string]string
// @Failure      409   {object}  map[string]string
// @Failure      500   {object}  map[string]string
// @Router       /users/confirm-email [post]
func (server *Server) confirmEmail(ctx *gin.Context) {
	var req ConfirmEmail

	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, errorResponse(err))
		return
	}

	//sanitize user email
	req.SanitizeConfirmEmailReq()

	user, err := server.store.GetUserByEmail(ctx, req.Email)
	if handleDBError(ctx, err,
		WithNotFoundMsg(ErrUserNotFound),
		WithLogArgs("confirmEmail: failed GetUserByEmail", "user_email", req.Email),
	) {
		return
	}

	//if user is already verified fail the request
	if user.EmailConfirmed {
		ctx.JSON(http.StatusConflict, errorMessage(ErrEmailAlreadyVerified))
		return
	}

	//2. Get valid OTP Code from db
	emailOTP, err := server.store.GetValidOTP(ctx, user.ID)
	if handleDBError(ctx, err,
		WithNotFoundMsg(ErrOTPNotFound),
		WithLogArgs("confirmEmail: failed GetValidOTP", "user_id", user.ID),
	) {
		return
	}

	//3. Compare the OTP hash to the given OTP
	match, err := auth.ConfirmOTP(req.EMAIL_OTP, emailOTP.CodeHash, server.config.TokenSymmetricKey)
	if err != nil {
		slog.Error("confirmEmail: failed confirmOTP", "err", err, "user_id", user.ID)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrInternalServer))
		return
	}
	if !match {
		_, err := server.store.IncrementOTPAttempts(ctx, emailOTP.ID)
		if err != nil {
			slog.Error("confirmEmail: failed IncrementOTPAttempts", "err", err, "user_id", user.ID)
			ctx.JSON(http.StatusInternalServerError, errorMessage(ErrInternalServer))
			return
		}

		//successful increment
		//slog.Error("failed to confirm OTP Code", "err", err, "user_id", user.ID)
		ctx.JSON(http.StatusForbidden, errorMessage(ErrOTPInvalid))
		return
	}
	//4. successful match, mark the OTP as used
	_, err = server.store.MarkOTPUsed(ctx, emailOTP.ID)
	if handleDBError(ctx, err, WithLogArgs("confirmEmail: failed MarkOTPUsed", "OTP_id", emailOTP.ID)) {
		return
	}

	//5. Mark user email as verified
	_, err = server.store.ConfirmEmail(ctx, user.ID)
	if err != nil {
		slog.Error("confirmEmail: failed ConfirmEmail", "err", err, "user_id", user.ID)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrInternalServer))
		return
	}

	ctx.JSON(http.StatusOK, ConfirmEmailResponse{
		Message: "email verified successfully",
		Email:   user.Email,
	})

}

type SendEmailOTP struct {
	Email string `json:"email" binding:"required,email"`
}

func (r *SendEmailOTP) SanitizeEmailOTP() {
	r.Email = strings.ToLower(r.Email)
}
type SendEmailOTPResp struct{
	Message string `json:"msg"`
	Email string `json:"email"`
}
// @Summary      Send OTP
// @Description  Resends the OTP verification code to the user's email.
// @Tags         Auth
// @Accept       json
// @Produce      json
// @Param        body  body  SendEmailOTP  true  "Email payload"
// @Success      200   {object}  SendEmailOTPResp
// @Failure      400   {object}  map[string]string
// @Failure      403   {object}  map[string]string
// @Failure      404   {object}  map[string]string
// @Failure      409   {object}  map[string]string
// @Failure      500   {object}  map[string]string
// @Router       /users/send-emailotp [post]
func (server *Server) sendEmailOTP(ctx *gin.Context) {
	var req SendEmailOTP
	if err := ctx.ShouldBindJSON(&req); err != nil {
		ctx.JSON(http.StatusBadRequest, errorResponse(err))
		return
	}

	//sanitize user email
	req.SanitizeEmailOTP()

	//2. Get the user by email
	user, err := server.store.GetUserByEmail(ctx, req.Email)
	if handleDBError(ctx, err,
		WithNotFoundMsg(ErrUserNotFound),
		WithLogArgs("sendEmailOTP: failed GetUserByEmail", "user_email", req.Email),
	) {
		return
	}

	if user.EmailConfirmed {
		ctx.JSON(http.StatusConflict, errorMessage(ErrEmailAlreadyVerified))
		return
	}

	//get a valid OTP if exists then the user cannot generate new OTP
	emailOTP, err := server.store.GetValidOTP(ctx, user.ID)
	if err != nil && !errors.Is(err, pgx.ErrNoRows) {
		if handleDBError(ctx, err, WithLogArgs("sendEmailOTP: failed GetValidOTP", "user_id", user.ID)) {
			return
		}
	}

	// valid OTP exists — enforce cooldown
	if err == nil && emailOTP.ExpiresAt.Time.After(time.Now()) {
		ctx.JSON(http.StatusForbidden, errorMessage(ErrOTPCooldown))
		return
	}
	//3. generate otp
	otpCode, err := auth.GenerateOTP(6)
	if err != nil {
		slog.Error("sendEmailOTP: failed GenerateOTP", "err", err, "user_id", user.ID)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrGeneratingOTP))
		return
	}

	//5. sign,Hash OTP and store in database
	hashedOTP, err := auth.SignOtpCode(otpCode, server.config.TokenSymmetricKey)
	if err != nil {
		slog.Error("sendEmailOTP: failed SignOtpCode",
			"err", err,
			"user_ID",user.ID,
		)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrInternalServer))
		return
	}
	_, err = server.store.CreateEmailOTP(ctx, db.CreateEmailOTPParams{
		UserID:   user.ID,
		CodeHash: hashedOTP,
		ExpiresAt: pgtype.Timestamptz{
			Time:  time.Now().Add(server.config.AcessTokenDuration),
			Valid: true,
		},
	})
	if handleDBError(ctx, err, WithLogArgs("sendEmailOT: failed CreateEmailOTP","user_id", user.ID)) {
		return
	}
	//4. send email otp to user email
	err = server.emailClient.SendEmailOTP(user.Email, otpCode)
	if err != nil {
		slog.Error("sendEmailOTP: failed SendEmailOTP", "err", err, "user_id", user.ID)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrSendingOTP))
		return
	}
	ctx.JSON(http.StatusOK, SendEmailOTPResp{
		Message: "OTP sent successfuly to your email",
		Email: user.Email,
	})

}

type LoginUserRequest struct {
	Email    string `json:"email" binding:"required,email"`
	Password string `json:"password" binding:"required,min=6"`
}

type LoginUserResponse struct {
	Username          string    `json:"username"`
	PasswordChangedAt time.Time `json:"password_changed_at"`
	LastLoginAt       time.Time `json:"last_login_at"`
	AccessToken       string    `json:"access_token"`
	RefreshToken      string    `json:"refresh_token"`
}

func (r *LoginUserRequest) sanitizeLoginUserReq() {
	r.Email = strings.ToLower(r.Email)
}
type EmailConfirmedResp struct{
		Message string `json:"msg"`
		Email string `json:"email"`
	}

// login user
// @Summary      Sign in
// @Description  Authenticates with email and password. Returns access + refresh tokens.
// @Tags         Auth
// @Accept       json
// @Produce      json
// @Param        body  body  LoginUserRequest  true  "Login payload"
// @Success      200   {object}  LoginUserResponse
// @Failure      400   {object}  map[string]string
// @Failure      401   {object}  map[string]string
// @Failure      404   {object}  map[string]string
// @Failure      500   {object}  map[string]string
// @Router       /users/signin [post]
func (server *Server) loginUser(ctx *gin.Context) {
	var req LoginUserRequest
	err := ctx.ShouldBindJSON(&req)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, errorResponse(err))
		return
	}

	//sanitize input
	req.sanitizeLoginUserReq()

	user, err := server.store.GetUserByEmail(ctx, req.Email)
	if handleDBError(ctx, err, WithNotFoundMsg(ErrUserNotFound),
		WithLogArgs("loginUser: failed GetUserByEmail","user_email", req.Email)) {
		return
	}

	//check if user email is verified
	if !user.EmailConfirmed{
		ctx.JSON(http.StatusForbidden,EmailConfirmedResp{
			Message: "email not confirmed!",
			Email: user.Email,
		})
		return
	}

	//check user password against saved db password
	err = auth.CheckPassword(user.PasswordHash, req.Password)
	if err != nil {
		ctx.JSON(http.StatusUnauthorized, errorMessage(ErrInvalidCredentials))
		return
	}

	//create the access token
	access_token, err := server.tokenMaker.CreateToken(user.Username, token.AccessTokenType, server.config.AcessTokenDuration)
	if err != nil {
		slog.Error("loginUser: failed CreateToken - accessToken", "err:", err, "user_email", req.Email)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrInternalServer))
		return
	}

	//create the refresh token signed with 1 hour and save to the database
	week := time.Hour * 24 * 7

	refresh_token, err := server.tokenMaker.CreateToken(user.Username, token.RefreshTokenType, week)
	if err != nil {
		slog.Error("loginUser: failed CreateToken - refreshToken", "err:", err, "user_id", user.ID)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrInternalServer))
		return
	}

	//create user session + refreshtoken in db

	savedSession, err := server.store.CreateSession(ctx, db.CreateSessionParams{
		UserID:       user.ID,
		RefreshToken: refresh_token,
		UserAgent: pgtype.Text{
			String: ctx.Request.UserAgent(),
			Valid:  true,
		},
		ClientIp: pgtype.Text{
			String: ctx.ClientIP(),
			Valid:  true,
		},
		ExpiresAt: pgtype.Timestamptz{
			Time:  time.Now().Add(week),
			Valid: true,
		},
	})
	if handleDBError(ctx, err, WithLogArgs("loginUser: failed CreateSession","user_id", user.ID)) {
		return
	}

	last_login := user.LastLoginAt
	err = server.store.UpdateLastLogin(ctx, user.ID)
	if handleDBError(ctx, err, WithLogArgs("loginUser: failed UpdateLastLogin","user_id", user.ID)) {
		return
	}

	resp := LoginUserResponse{
		Username:          user.Username,
		PasswordChangedAt: user.PasswordUpdatedAt.Time,
		LastLoginAt:       last_login.Time,
		AccessToken:       access_token,
		RefreshToken:      savedSession.RefreshToken,
	}

	ctx.JSON(http.StatusOK, resp)

}

type RefreshTokenRequest struct {
	RefreshToken string `json:"refresh_token" binding:"required"`
}


type RefreshTokenResponse struct {
	Message string `json:"message"`
	AccessToken   string `json:"access_token"`
}

// @Summary      Refresh token
// @Description  Issues a new access token using a valid refresh token.
// @Tags         Auth
// @Accept       json
// @Produce      json
// @Param        body  body  RefreshTokenRequest  true  "Refresh token payload"
// @Success      200   {object}  RefreshTokenResponse
// @Failure      400   {object}  map[string]string
// @Failure      401   {object}  map[string]string
// @Failure      500   {object}  map[string]string
// @Router       /users/refresh-token [post]
func (server *Server) refreshToken(ctx *gin.Context) {
	var req RefreshTokenRequest

	err := ctx.ShouldBindJSON(&req)
	if err != nil {
		ctx.JSON(http.StatusBadRequest, errorResponse(err))
		return
	}

	//verify the refresh token and get the payload
	payload, err := server.tokenMaker.VerifyToken(req.RefreshToken, token.RefreshTokenType)
	if err != nil {
		ctx.JSON(http.StatusUnauthorized, errorMessage(ErrInvalidToken))
		return
	}


    // 2. check session in DB — this is what JWT alone can't do
    session, err := server.store.GetSessionByRefreshToken(ctx, req.RefreshToken)
    if handleDBError(ctx, err, WithNotFoundMsg(ErrSessionNotFound)) {
        return
    }

    // 3. check not revoked
    if session.IsRevoked {
        ctx.JSON(http.StatusUnauthorized, errorMessage(ErrSessionRevoked))
        return
    }

    // 4. check not expired
    if session.ExpiresAt.Time.Before(time.Now()) {
        ctx.JSON(http.StatusUnauthorized, errorMessage(ErrSessionExpired))
        return
    }

	//refreshtoken is valid issue new access token
	accessToken, err := server.tokenMaker.CreateToken(payload.Subject, token.AccessTokenType, server.config.AcessTokenDuration)
	if err != nil {
		slog.Error("refreshToken: failed CreateToken - accessToken","user_id",payload.ID)
		ctx.JSON(http.StatusInternalServerError, errorMessage(ErrInvalidToken))
		return
	}

	resp := RefreshTokenResponse{
		Message: "successfully created accessToken",
		AccessToken: accessToken,
	}

	ctx.JSON(http.StatusOK, resp)
}